fix playlist sort button ui
